+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The version format is MCMAJOR.MCMINOR.MCPATCH-MAJOR.MINOR.PATCH, eg, 1.18.1-1.2.
 
 ### Changed
 - Detectors no longer conduct redstone power
+- Made nightshale textures slightly nicer
 
 ## [1.18.1-0.2.0-alpha]
 ### Added