+++ CHANGELOG.md
@@ -14,7 +14,8 @@ The version format is MCMAJOR.MCMINOR.MCPATCH-MAJOR.MINOR.PATCH, eg, 1.18.1-1.2.
 ### Changed
 - Detectors no longer conduct redstone power
   - This behavior matches that of vanilla observers
-- Made nightshale textures slightly nicer
+- Made nightshale textures slightly nicer™
+  - They're now less saturated and slightly more purple
 - Nightshale now uses deepslate sounds
 
 ## [1.18.1-0.2.0-alpha]